requirements.txt
imaplib2
ConfigParser
email
virustotal-api==1.1.2